refactor: split library type screen
